Move deletion related logic as much as possible to tenant/delete.rs
